Refactor and fix some style issues
