+++ src/components/Layout.tsx
@@ -1,8 +1,12 @@
 import { ReactNode } from "react";
+import Navbar from "./Navbar";
 
 const Layout = ({ children }: { children: ReactNode }) => {
   return (
-    <main className="-mt-10 mx-8 md:mx-12 lg:mx-[10%]">{children}</main>
-  )
-}
-export default Layout
+    <>
+      <Navbar />
+      <main className="-mt-10 mx-8 md:mx-12 lg:mx-[10%]">{children}</main>
+    </>
+  );
+};
+export default Layout;
