+++ src/components/details/ApplyButtonComponent.tsx
