+++ src/components/details/JobTitleCard.tsx
