+++ src/components/details/JobTitleDescriptions.tsx
