+++ src/components/details/JobTitleFooter.tsx
